perf: Made features register themselves, as suggested by schneefux feat(playlist): Initial
Playlist feature now plays music using Lavaplayer in the playlist voice channel configured for each server.
Music is chosen in a random order but synchronised across all voice channels the bot is in.
However, for each server the audio is playing in, the speed of the song is multiplied (this is a bug).
I assume this is because one server cannot listen to the same segment of the audio another server has already listened to due to how I have currently set up the buffer. One server will hear the first 50ms but will skip the 50ms after that because that's what another server will have heard first.